ci: re-enable chrome mobile tests on android (#32447)
With 57c76c36c275e45a0fb1ea2983f99278a6bb15d3, we temporarily disabled
the chrome mobile tests on Android because the launched browsers weren't
able to load the karma page through the tunnel.

The Saucelabs support responded to a ticket that we opened and said that
there were issues with those browsers which are now resolved.

PR Close #32447