fix(forms): disabled input acceptance member not properly applied (#34502)
With d3a2ce38b05255f431d6b0fb0fcb61f7b1ac17f6 we intended to expand
the input type of the `disabled` input of the `NgModel` directive.
Read more about the reason for this in the actual commit message.

Currently though, the acceptance coercion member does not have any
effect. This is because the acceptance member needs to refer to the
actual input property name, and not to the public input name.
`disabled` corresponds to the `isDisabled` property.

PR Close #34502