refactor(language-service): rename expression completions method (#34518)
This commit renames `addAttributeValuesToCompletions`, which generates
expression completions and is not exclusive to processing attributes, to
`processExpressionCompletions`. Also removes the expression completion
logic in `visitBoundText` for a call to `processExpressionCompletions`.

The conditional branch in `visitBoundText` is also removed. This branch
was added in one of the first commits to the language service
(48402d6ae532805e12bb2b8bbf6b709bc3d82341) and appears to be
unnecessary, as the expression AST is constructed from the template
position anyway.

PR Close #34518