Revert "refactor(compiler): refactor and export Bazel host creation. (#34686)" (#35063)
This reverts commit 4d551c0a21d99d4e71b970f9f29a399ab361f58f.

The intention of this commit was for a consumer of the `compile` function to
pass the `bazelHost` it returns into future invocations, reusing the
`FileCache` between them. However, first-party ngc_wrapped does not do this,
which caused a performance regression as the `FileCache` was no longer
shared between compilations.

PR Close #35063