fix(ngcc): do not warn if `paths` mapping does not exist (#36525)
In bf6afe749158ebad935b138534f714abb6f907e0 the `getBasePaths()`
function was changed to log a warning if a `basePath()` computed from
the `paths` mappings did not exist. It turns out this is a common and
accepted scenario, so we should not log warnings in this case.

Fixes #36518

PR Close #36525