ci: increase payload size limits for integration tests (#37123)
This commit updates payload size limits that are triggering errors after merging abf488e. That commit seems to contribute to the payload size increase, but all checks were "green" for the original PR (#35889), so it looks like it's an accumulated payload size increase from multiple changes. The goal of this commit is to bring the master branch back to "green" state.

PR Close #37123